client: Ok, ok, semicolons are cool.
Yes, there's some <http://jsdrama.com> around and people have opinions
about it... I have one of my own, too, but making it easier for others
to read my code is what sold me.
So, yay semicolons.

(But why is it allowed to omit the semicolons in the first place? I
don't get it!)